requirements.txt
pycryptodomex==3.9.8
django>=1.10